chore!: upgrade to cdktf 0.18.0 (#142)
This PR initiates the upgrade of CDKTF from version `0.17.0` to version
`0.18.0`.
Unfortunately, not everything can be automated, and the following steps
need to be completed manually:
 
- [ ] Update `@cdktf/provider-tfe` to a version compatible with
`cdktf@0.18.0`
[here](https://github.com/cdktf/cdktf-multi-stack-tfe/blob/d354d32382b3aa20cdc28c332131d04b730623c0/.projenrc.ts#L36)
 
Please checkout this PR, complete the above steps, push the changes to
this branch, and then mark this PR as ready for review to complete the
upgrade. Thanks!

---------

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: Nara Kasbergen Kwon <855115+xiehan@users.noreply.github.com>
Co-authored-by: Nara Kasbergen <nara.kasbergen@gmail.com>