chore!: upgrade to cdktf 0.19.0 (#158)
This PR initiates the upgrade of CDKTF from version `0.18.0` to version
`0.19.0`.
Unfortunately, not everything can be automated, and the following steps
need to be completed manually:
 
- [x] Update `@cdktf/provider-tfe` to a version compatible with
`cdktf@0.19.0`
[here](https://github.com/cdktf/cdktf-multi-stack-tfe/blob/2aa8e48abe9dd9ccc6b3714403a2eea72eaf00b5/.projenrc.ts#L65).
Look up the version
[here](https://github.com/cdktf/cdktf-provider-tfe/releases/).
- [x] Run `npx projen`
 
Please checkout this PR, complete the above steps, push the changes to
this branch, and then mark this PR as ready for review to complete the
upgrade. Thanks!

---------

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: Daniel Schmidt <danielmschmidt92@gmail.com>